+++ src/modules/types/server_interaction/webservice_search/view.js
@@ -189,6 +189,11 @@ define(['modules/default/defaultview'], function (Default) {
         onActionReceive: {
             doSearch: function () {
                 this.module.controller.doSearch();
+            },
+            buttonColor: function (newColor) {
+                if (this.buttonInst) {
+                    this.buttonInst.setColor(newColor);
+                }
             }
         },