+++ History.md
@@ -4,6 +4,7 @@ HEAD
 * button_action : add option to disable toggle
 * datas : fix several issues with DataObject
 * add API.cache
+* webservice_search : add action to change button color
 
 2.8.0 / 2014-11-10
 ===================